+++ test/controllers/settings_controller_test.rb
@@ -18,6 +18,15 @@ class SettingsControllerTest < ActionDispatch::IntegrationTest
     assert_select "h2", text: "Feeder Invites", count: 0
   end
 
+  test "should link to settings sections" do
+    sign_in_user
+    get settings_url
+    assert_response :success
+    assert_select "a[href=?]", access_tokens_path
+    assert_select "a[href=?]", ai_credentials_path
+    assert_select "a[href=?]", invites_path
+  end
+
   test "should redirect to login when not authenticated" do
     get settings_url
     assert_redirected_to new_session_path